backend/pyproject.toml
[tool.poetry]
name = "teamai-backend"
version = "0.1.0"
description = "TeamAI Backend - Virtual AI Workforce Platform"
authors = ["dlai-sd <your@email.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Database
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"

# Redis
redis = "^5.0.1"
hiredis = "^2.3.2"

# Azure SDK
azure-identity = "^1.15.0"
azure-keyvault-secrets = "^4.7.0"
azure-storage-blob = "^12.19.0"
azure-functions = "^1.18.0"
azure-functions-durable = "^1.2.9"

# LLM Providers
openai = "^1.10.0"
groq = "^0.4.1"
langchain = "^0.1.0"
langchain-community = "^0.0.16"
langgraph = "^0.0.20"

# HTTP Clients & Web Scraping
httpx = "^0.26.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"
pypdf2 = "^3.0.1"

# Auth & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"

# Utilities
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.12.1"
pylint = "^3.0.3"
mypy = "^1.8.0"
ipython = "^8.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
